+++ queries/javascript/rainbow-tags-react.scm
@@ -34,8 +34,14 @@
 
 ;;; React.js support
 (jsx_element
-  open_tag: (jsx_opening_element) @opening
-  close_tag: (jsx_closing_element) @closing) @container
+  open_tag: (jsx_opening_element
+              name: (identifier) @opening)
+  close_tag: (jsx_closing_element
+               name: (identifier) @closing)) @container
+
+(jsx_self_closing_element
+  name: (identifier) @opening
+  "/" @closing) @container
 
 (jsx_expression
   "{" @opening